+++ ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: usolak <usolak@student.42istanbul.com.t    +#+  +:+       +#+        */
+/*   By: usolak <usolak@student.42istanbul.com.tr>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/26 11:33:32 by usolak            #+#    #+#             */
-/*   Updated: 2026/02/02 22:21:10 by usolak           ###   ########.fr       */
+/*   Updated: 2026/02/03 14:12:50 by usolak           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,6 +22,6 @@ int	ft_print_str(char *str);
 int	ft_print_pointer(unsigned long nb);
 int	ft_print_int(int n);
 int	ft_print_unsigned(unsigned int nb);
-int	ft_print_hex(unsigned int nb, char small_x);
+int	ft_print_hex(unsigned long nb, char small_x);
 
 #endif
